backend/requirements.txt
# django packages
Django==3.2.9
django-lifecycle>=0.9.1
djangorestframework>=3.12.4
djangorestframework-simplejwt

Pillow>=8.4.0  # using for upload images


# django channels
channels>=3.0.4
channels-redis>=3.3.1

# cache
django-redis>=5.0.0

# server runner
uvicorn==0.15.0

# ws uvicorn dependencies
httptools>=0.2.0
python-dotenv>=0.19.0
uvloop>=0.16.0
watchgod>=0.7
websockets>=10.0

# database
psycopg2-binary>=2.9.1

# celery
Celery>=5.1.2
django-celery-results>=2.2.0


# swagger api docs
drf-spectacular>=0.19.0

faker

